add more tests and move naive_einsum to funsor.testing
