+++ funsor/testing.py
@@ -9,8 +9,9 @@ import torch
 from six import integer_types
 from six.moves import reduce
 
+import funsor.ops as ops
 from funsor.domains import bint
-from funsor.terms import Funsor
+from funsor.terms import Binary, Funsor
 from funsor.torch import Tensor
 
 
@@ -95,3 +96,27 @@ def random_tensor(dtype, shape):
         return torch.randn(shape)
     else:
         raise ValueError('unknown dtype: {}'.format(repr(dtype)))
+
+
+def naive_einsum(eqn, *terms, **kwargs):
+    backend = kwargs.pop('backend', 'torch')
+    if backend == 'torch':
+        sum_op, prod_op = ops.add, ops.mul
+    elif backend == 'pyro.ops.einsum.torch_log':
+        sum_op, prod_op = ops.logaddexp, ops.add
+    else:
+        raise ValueError("{} backend not implemented".format(backend))
+
+    assert isinstance(eqn, str)
+    assert all(isinstance(term, Funsor) for term in terms)
+    inputs, output = eqn.split('->')
+    assert len(output.split(',')) == 1
+    input_dims = frozenset(d for inp in inputs.split(',') for d in inp)
+    output_dims = frozenset(d for d in output)
+    reduce_dims = tuple(d for d in input_dims - output_dims)
+    prod = terms[0]
+    for term in terms[1:]:
+        prod = Binary(prod_op, prod, term)
+    for reduce_dim in reduce_dims:
+        prod = prod.reduce(sum_op, reduce_dim)
+    return prod
